+++ packages/core/src/web/index.ts
@@ -1,4 +1,5 @@
 export * from "./appInfo";
+export * from "./dashboardMetricColors";
 export * from "./improvementTimeline";
 export * from "./jobSchemas";
 export * from "./jobUpload";